requirements.txt
mysql-connector-python
matplotlib
seaborn
pandas
tk